+++ x-objects/lib/x/objects/api/lookups.rb
@@ -1,6 +1,7 @@
 require_relative "lookups/communities"
 require_relative "lookups/direct_messages"
 require_relative "lookups/lists"
+require_relative "lookups/media"
 require_relative "lookups/posts"
 require_relative "lookups/spaces"
 require_relative "lookups/users"
@@ -14,6 +15,7 @@ module X
         include Users
         include Posts
         include Lists
+        include Media
         include Spaces
         include Communities
         include DirectMessages