+++ x-objects/test/x/objects/cursor_stub_hydration_test.rb
@@ -67,8 +67,8 @@ module X
     end
 
     def test_only_resources_with_a_batch_lookup_are_batchable
-      assert_equal [true, true, true], [User, Post, Space].map(&:batchable?)
-      assert_equal [false, false, false, false], [List, Community, DirectMessage, Media].map(&:batchable?)
+      assert_equal [true, true, true, true], [User, Post, Space, Media].map(&:batchable?)
+      assert_equal [false, false, false, false], [List, Community, DirectMessage, Poll].map(&:batchable?)
     end
 
     def test_refreshing_a_stub_of_a_page_looks_it_up_again_on_its_own