+++ x-objects/test/x/objects/cursor_token_test.rb
@@ -70,9 +70,9 @@ module X
     end
 
     def test_ids_without_a_fields_parameter
-      error = assert_raises(UnsupportedOperation) { Cursor.new(Media, "media", client: @client).ids }
+      error = assert_raises(UnsupportedOperation) { Cursor.new(Poll, "polls", client: @client).ids }
 
-      assert_equal "X::Media has no fields parameter", error.message
+      assert_equal "X::Poll has no fields parameter", error.message
       assert_empty @client.requests
     end